Convert search bar to its own component and add initial search keyword functionality on location
