+++ src/features/restaurants/components/Search.js
@@ -1,0 +1,29 @@
+import React, { useContext, useState } from 'react';
+import { View } from 'react-native';
+import { Searchbar } from 'react-native-paper';
+import styled from 'styled-components';
+
+import { LocationContext } from '../../../services/location/LocationContext';
+
+const SearchContainer = styled(View)`
+  padding: ${(props) => props.theme.space[3]};
+  background-color: #33ff90;
+`;
+
+export const Search = () => {
+  const { keyword, search } = useContext(LocationContext);
+  const [searchKeyword, setSearchKeyword] = useState(keyword);
+
+  return (
+    <SearchContainer>
+      <Searchbar
+        placeholder="Search for a location"
+        value={searchKeyword}
+        onSubmitEditing={() => search(searchKeyword)}
+        onChangeText={(text) => {
+          setSearchKeyword(text);
+        }}
+      />
+    </SearchContainer>
+  );
+};
