+++ src/features/restaurants/screens/RestaurantsScreen.js
@@ -1,18 +1,14 @@
 import React, { useContext } from 'react';
 import { FlatList, View } from 'react-native';
-import { Searchbar, ActivityIndicator, Colors } from 'react-native-paper';
+import { ActivityIndicator, Colors } from 'react-native-paper';
 import styled from 'styled-components/native';
 
 import { SafeArea } from '../../../components/SafeArea/SafeArea';
 import { Spacer } from '../../../components/Spacer/Spacer';
+import { Search } from '../components/Search';
 import { RestaurantInfoCard } from '../components/RestaurantInfoCard';
 
 import { RestaurantsContext } from '../../../services/restaurants/RestaurantsContext';
-
-const SearchContainer = styled(View)`
-  padding: ${(props) => props.theme.space[3]};
-  background-color: #33ff90;
-`;
 
 const RestaurantList = styled(FlatList).attrs({
   contentContainerStyle: {
@@ -40,9 +36,7 @@ export const RestaurantsScreen = () => {
           <Loading size={50} animating={true} color={Colors.blue300} />
         </LoadingContainer>
       )}
-      <SearchContainer>
-        <Searchbar placeholder="Search" />
-      </SearchContainer>
+      <Search />
       <RestaurantList
         data={restaurants}
         renderItem={({ item }) => {
